+++ hutool-crypto/src/main/java/cn/hutool/crypto/asymmetric/AsymmetricAlgorithm.java
@@ -8,8 +8,13 @@ package cn.hutool.crypto.asymmetric;
  *
  */
 public enum AsymmetricAlgorithm {
-	/** RSA算法，此算法用了默认补位方式为RSA/ECB/PKCS1Padding，不同JDK默认的补位方式可能不同，如Android默认是RSA/None/NoPadding */
-	RSA("RSA/ECB/PKCS1Padding"), 
+	/** RSA算法 */
+	RSA("RSA"), 
+	/** RSA算法，此算法用了默认补位方式为RSA/ECB/PKCS1Padding */
+	RSA_ECB_PKCS1("RSA/ECB/PKCS1Padding"), 
+	/** RSA算法，此算法用了RSA/None/NoPadding */
+	RSA_None("RSA/None/NoPadding"), 
+	/** EC算法 */
 	EC("EC");
 
 	private String value;